refactor(btc-staking): move full validation logic to separate module
Moves the validation logic to a new `full.rs` module and uses the `full-validation` feature flag to conditionally compile it.

This helps to keep the main validation module clean and makes it easier to manage the validation logic.
